requirements.txt
flake8==4.0.1
Pillow==9.0.0